requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
mysqlclient==2.2.1
python-dotenv==1.0.0

# AI/ML Stack
langchain==0.3.13
langgraph==0.2.59
langchain-openai==0.2.14
langchain-anthropic==0.3.7
openai==1.59.6
anthropic>=0.45.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.3

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1

# Utilities
python-dateutil==2.8.2
Faker==22.0.0

# Development
ipython==8.19.0

